+++ wakamiti/wakamiti-core/src/test/java/es/iti/wakamiti/test/gherkin/TestFailureStep.java
@@ -10,19 +10,24 @@
 package es.iti.wakamiti.test.gherkin;
 
 
+import es.iti.wakamiti.api.event.Event;
+import es.iti.wakamiti.api.plan.*;
 import imconfig.Configuration;
 import es.iti.wakamiti.api.WakamitiConfiguration;
-import es.iti.wakamiti.api.plan.PlanNode;
-import es.iti.wakamiti.api.plan.PlanNodeSnapshot;
-import es.iti.wakamiti.api.plan.PlanSerializer;
 import es.iti.wakamiti.core.JsonPlanSerializer;
 import es.iti.wakamiti.core.Wakamiti;
 import es.iti.wakamiti.core.gherkin.GherkinResourceType;
 import org.junit.Test;
 
+import java.io.FileWriter;
 import java.io.IOException;
+import java.io.Writer;
+import java.nio.file.Path;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -54,6 +59,21 @@ public class TestFailureStep {
         String serial = serializer.serialize(snapshot);
         System.out.println(serial);
         assertThat(serial).contains("\"errorClassifier\" : \"test\"");
+
+
+        List<PlanNode> testCases = plan
+            .descendants()
+            .filter(node -> node.nodeType() == NodeType.TEST_CASE)
+            .filter(node -> node.result().filter(it -> it == Result.FAILED).isPresent())
+            .collect(Collectors.toList());
+
+        for (PlanNode testCase : testCases) {
+            String testCaseSerial = serializer.serialize(new PlanNodeSnapshot(testCase).withoutChildren());
+            System.out.println("-----------------------------------------\n\n\n");
+            System.out.println(testCaseSerial);
+            assertThat(testCaseSerial).contains("\"errorClassifier\" : \"test\"");
+        }
+
 
     }