+++ wakamiti/wakamiti-api/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [2.0.1] - 2023-06-22
+
+
+### Fixed
+
+- Output field `errorClassifier` was not included when `outputFilePerTestCase = true`
+
+
+
+
 ## [2.0.0] - 2023-05-22